+++ CHANGELOG.md
@@ -4,9 +4,13 @@ This file contains highlights of what changes on each version of the StageXL_Par
 package. This file is normally updated whenever we push a new version to pub.
 
 #### Pub version 1.0.0
-* Raised dependency to StageXL 2.0.0
-* Update Dart SDK restraint to >= 2.12.0
 * Upgrade to null safety
+* Update Dart SDK restraint to >= 2.12.0
+* Raised dependency to StageXL 2.0.0
+
+#### Pub version 0.4.1
+  * Ready for Dart 2.4
+  * Raised dependency to StageXL 1.4.1
 
 #### Pub version 0.4.0
   * Raised dependency to StageXL 1.4.0